Move the rule to the new package name
