+++ src/test/java/hudson/security/LDAPEmbeddedTest.java
@@ -30,6 +30,9 @@ import hudson.tasks.Mailer;
 import hudson.util.Secret;
 import jenkins.model.IdStrategy;
 import jenkins.security.plugins.ldap.FromGroupSearchLDAPGroupMembershipStrategy;
+import jenkins.security.plugins.ldap.LDAPConfiguration;
+import jenkins.security.plugins.ldap.LDAPRule;
+import jenkins.security.plugins.ldap.LDAPSchema;
 import org.acegisecurity.userdetails.ldap.LdapUserDetails;
 import org.junit.Rule;
 import org.junit.Test;
